Refactor routes module to split it into several different files
