+++ hushline/routes/message.py
@@ -1,0 +1,118 @@
+from datetime import UTC, datetime
+
+from flask import (
+    Flask,
+    abort,
+    current_app,
+    flash,
+    redirect,
+    render_template,
+    session,
+    url_for,
+)
+from werkzeug.wrappers.response import Response
+
+from hushline.auth import authentication_required
+from hushline.db import db
+from hushline.forms import DeleteMessageForm, UpdateMessageStatusForm
+from hushline.model import (
+    Message,
+    User,
+    Username,
+)
+
+
+def register_message_routes(app: Flask) -> None:
+    @app.route("/message/<int:id>")
+    @authentication_required
+    def message(id: int) -> str:
+        msg = db.session.scalars(
+            db.select(Message)
+            .join(Username)
+            .filter(Username.user_id == session["user_id"], Message.id == id)
+        ).one_or_none()
+
+        if not msg:
+            abort(404)
+
+        update_status_form = UpdateMessageStatusForm(data={"status": msg.status.value})
+        delete_message_form = DeleteMessageForm()
+
+        return render_template(
+            "message.html",
+            message=msg,
+            update_status_form=update_status_form,
+            delete_message_form=delete_message_form,
+        )
+
+    @app.route("/reply/<slug>")
+    def message_reply(slug: str) -> str:
+        msg = db.session.scalars(db.select(Message).filter_by(reply_slug=slug)).one_or_none()
+        if msg is None:
+            abort(404)
+
+        return render_template("reply.html", message=msg)
+
+    @app.route("/message/<int:id>/delete", methods=["POST"])
+    @authentication_required
+    def delete_message(id: int) -> Response:
+        user = db.session.scalars(db.select(User).filter_by(id=session["user_id"])).one()
+
+        row_count = db.session.execute(
+            db.delete(Message).where(
+                Message.id == id,
+                Message.username_id.in_(
+                    db.select(Username.id).select_from(Username).filter(Username.user_id == user.id)
+                ),
+            )
+        ).rowcount
+        match row_count:
+            case 1:
+                db.session.commit()
+                flash("🗑️ Message deleted successfully.")
+            case 0:
+                db.session.rollback()
+                flash("⛔️ Message not found.")
+            case _:
+                db.session.rollback()
+                current_app.logger.error(
+                    f"Multiple messages would have been deleted. Message.id={id} User.id={user.id}"
+                )
+                flash("Internal server error. Message not deleted.")
+
+        return redirect(url_for("inbox"))
+
+    @app.route("/message/<int:id>/status", methods=["POST"])
+    @authentication_required
+    def set_message_status(id: int) -> Response:
+        user = db.session.scalars(db.select(User).filter_by(id=session["user_id"])).one()
+
+        form = UpdateMessageStatusForm()
+        if not form.validate():
+            flash(f"Invalid status: {form.status.data}")
+            return redirect(url_for("message", id=id))
+
+        row_count = db.session.execute(
+            db.update(Message)
+            .where(
+                Message.id == id,
+                Message.username_id.in_(
+                    db.select(Username.id).select_from(Username).filter(Username.user_id == user.id)
+                ),
+            )
+            .values(status=form.status.data, status_changed_at=datetime.now(UTC))
+        ).rowcount
+        match row_count:
+            case 1:
+                db.session.commit()
+                flash("👍 Message status updated.")
+            case 0:
+                db.session.rollback()
+                flash("⛔️ Message not found.")
+            case _:
+                db.session.rollback()
+                current_app.logger.error(
+                    f"Multiple messages would have been updated. Message.id={id} User.id={user.id}"
+                )
+                flash("Internal server error. Message not updated.")
+        return redirect(url_for("message", id=id))
